We've decided to get rid of shmemipc because the two 16K blocks outside of main RAM cannot, by design, be used for what we want to use them for. Shmemipc is inherently broken by design.
This commit does not remove shmemipc completely, but it does
a first step. The wifi driver now uses the FIFO and memory in the 4MB main
RAM to send and recieve packets. It still uses shmemipc for AP scanning
and statistics, but that will be changed soon as well.

The firmware dumper also needs to be adjusted before we can throw
shmemipc out completely.

This commit also modifies the interface of the FIFO driver. FIFO commands
are now encoded differently, in order to have more bits left for wifi
commands. See the comments in the code for details.

Some wifi fifo commands that were related have been combined
into single commands.

Some "magic numbers" in the FIFO driver have been replaced by named
constants.