Use proper CFLAGS when compiling OpenSSL.
I don't know why I came up with the idea of renaming the CFLAG
variable in this script to CFLAGS, it seemed like the right
thing to do at the time :-)

Well, back then the Makefile was static so proper CFLAGS
were hardcoded in it anyway. But now that the Makefile is
created dynamically by the configure script this bit us in
the buttocks. I apologise for being so stupid, in hindsight.

Also, fix the dslinux platform description wrt asm code and -fPIC.