Disable sed again.
The build keeps jumping into subdirectories that fail to build.
For example, docs/ - building docs fails because help2man tries
to run the cross-compiled sed binary on the host system.

Trying to narrow down the list of subdirectories to build
in user/sed/Makefile fails, because that file is sometimes
regenerated for some reason.

So I am disabling this until someone fixes this port so
it builds reliably. Preferrably make this port use its
configure script properly please.