Apply telnetd patch by Sonny Jim:
I've just checked out a clean svn for the first time in a while and I
noticed that telnetd still doesn't work on the DLDI build.  Here's a
patch that fixes it to run through inetd which works fine.  Also it
creates some entries for /etc/services, which it didn't before